requirements.txt
confluent-kafka==2.9.0
python-dotenv==1.1.0
